Create ChessGame and move initialize and move from Board
