+++ src/main/java/kr/codesquad/freddie/chess/Main.java
@@ -2,6 +2,7 @@ package kr.codesquad.freddie.chess;
 
 import kr.codesquad.freddie.chess.board.Board;
 import kr.codesquad.freddie.chess.command.Command;
+import kr.codesquad.freddie.chess.game.ChessGame;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -10,14 +11,14 @@ import java.io.InputStreamReader;
 public class Main {
     public static void main(String[] args) throws IOException {
         try (BufferedReader br = new BufferedReader(new InputStreamReader(System.in))) {
-            Board board = new Board();
+            ChessGame chessGame = ChessGame.create();
 
             for (String command = br.readLine();
                  !command.equals("end");
                  command = br.readLine()) {
                 try {
-                    Command.create(command).execute(board);
-                    System.out.println(board.getRepresentation());
+                    Command.create(command).execute(chessGame);
+                    System.out.println(chessGame.getBoard().getRepresentation());
                 } catch (IllegalArgumentException e) {
                     System.out.println(e.getMessage());
                 }
